Refactor : 디렉토리명 도메인에 맞게 변경
