+++ src/main/java/com/eduquest/backend/application/community/service/AnswerService.java
@@ -13,8 +13,8 @@ import com.eduquest.backend.domain.community.service.AnswerCommandService;
 import com.eduquest.backend.domain.community.service.AnswerQueryService;
 import com.eduquest.backend.domain.community.service.QuestionCommandService;
 import com.eduquest.backend.domain.community.service.QuestionQueryService;
-import com.eduquest.backend.domain.member.model.Member;
-import com.eduquest.backend.domain.member.service.MemberQueryService;
+import com.eduquest.backend.domain.identity.model.Member;
+import com.eduquest.backend.domain.identity.service.MemberQueryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
