+++ src/main/java/com/eduquest/backend/application/identity/service/AdminUserService.java
@@ -1,9 +1,9 @@
 package com.eduquest.backend.application.identity.service;
 
-import com.eduquest.backend.domain.member.event.AccountLockEvent;
-import com.eduquest.backend.domain.member.model.Member;
-import com.eduquest.backend.domain.member.service.MemberCommandService;
-import com.eduquest.backend.domain.member.service.MemberQueryService;
+import com.eduquest.backend.domain.identity.event.AccountLockEvent;
+import com.eduquest.backend.domain.identity.model.Member;
+import com.eduquest.backend.domain.identity.service.MemberCommandService;
+import com.eduquest.backend.domain.identity.service.MemberQueryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
