+++ src/main/java/com/eduquest/backend/application/identity/service/SignUpService.java
@@ -8,11 +8,11 @@ import com.eduquest.backend.domain.file.model.File;
 import com.eduquest.backend.domain.file.model.StorageType;
 import com.eduquest.backend.domain.file.service.FileCommandService;
 import com.eduquest.backend.domain.file.service.FileQueryService;
-import com.eduquest.backend.domain.member.component.CustomPasswordEncoder;
-import com.eduquest.backend.domain.member.model.Member;
-import com.eduquest.backend.domain.member.model.enums.RoleType;
-import com.eduquest.backend.domain.member.service.MemberCommandService;
-import com.eduquest.backend.domain.member.service.MemberQueryService;
+import com.eduquest.backend.domain.identity.component.CustomPasswordEncoder;
+import com.eduquest.backend.domain.identity.model.Member;
+import com.eduquest.backend.domain.identity.model.enums.RoleType;
+import com.eduquest.backend.domain.identity.service.MemberCommandService;
+import com.eduquest.backend.domain.identity.service.MemberQueryService;
 import com.eduquest.backend.domain.reward.event.GrantPointEvent;
 import com.eduquest.backend.domain.reward.service.WalletCommandService;
 import com.eduquest.backend.infrastructure.s3.client.EduQuestS3Client;
