+++ src/main/java/com/eduquest/backend/application/identity/service/UserProfileService.java
@@ -9,11 +9,11 @@ import com.eduquest.backend.domain.file.model.File;
 import com.eduquest.backend.domain.file.model.StorageType;
 import com.eduquest.backend.domain.file.service.FileCommandService;
 import com.eduquest.backend.domain.file.service.FileQueryService;
-import com.eduquest.backend.domain.member.component.CustomPasswordEncoder;
-import com.eduquest.backend.domain.member.dto.MemberQuery;
-import com.eduquest.backend.domain.member.model.Member;
-import com.eduquest.backend.domain.member.service.MemberCommandService;
-import com.eduquest.backend.domain.member.service.MemberQueryService;
+import com.eduquest.backend.domain.identity.component.CustomPasswordEncoder;
+import com.eduquest.backend.domain.identity.dto.MemberQuery;
+import com.eduquest.backend.domain.identity.model.Member;
+import com.eduquest.backend.domain.identity.service.MemberCommandService;
+import com.eduquest.backend.domain.identity.service.MemberQueryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
