+++ src/main/java/com/eduquest/backend/application/note/service/NoteService.java
@@ -6,8 +6,8 @@ import com.eduquest.backend.application.note.dto.NoteListResult;
 import com.eduquest.backend.application.note.dto.UpdateNoteCommand;
 import com.eduquest.backend.application.note.exception.NoteErrorCode;
 import com.eduquest.backend.common.exception.EduQuestException;
-import com.eduquest.backend.domain.member.model.Member;
-import com.eduquest.backend.domain.member.service.MemberQueryService;
+import com.eduquest.backend.domain.identity.model.Member;
+import com.eduquest.backend.domain.identity.service.MemberQueryService;
 import com.eduquest.backend.domain.note.dto.NoteQuery;
 import com.eduquest.backend.domain.note.model.Note;
 import com.eduquest.backend.domain.note.service.NoteCommandService;
