+++ src/main/java/com/eduquest/backend/application/progress/service/ProgressService.java
@@ -4,7 +4,7 @@ import com.eduquest.backend.application.progress.dto.ProgressDto;
 import com.eduquest.backend.domain.learning.dto.ProblemQuery;
 import com.eduquest.backend.domain.learning.service.ProblemQueryService;
 import com.eduquest.backend.domain.learning.service.StageQueryService;
-import com.eduquest.backend.domain.member.service.MemberQueryService;
+import com.eduquest.backend.domain.identity.service.MemberQueryService;
 import com.eduquest.backend.domain.progress.dto.ProgressQuery;
 import com.eduquest.backend.domain.submission.model.Evaluation;
 import com.eduquest.backend.domain.submission.model.Submission;
