+++ src/main/java/com/eduquest/backend/application/submission/listener/SubmissionEventListener.java
@@ -1,8 +1,8 @@
 package com.eduquest.backend.application.submission.listener;
 
 import com.eduquest.backend.domain.learning.service.StageQueryService;
-import com.eduquest.backend.domain.member.model.Member;
-import com.eduquest.backend.domain.member.service.MemberQueryService;
+import com.eduquest.backend.domain.identity.model.Member;
+import com.eduquest.backend.domain.identity.service.MemberQueryService;
 import com.eduquest.backend.domain.reward.event.GrantPointEvent;
 import com.eduquest.backend.domain.submission.event.SubmissionEvaluatedEvent;
 import lombok.RequiredArgsConstructor;
