+++ src/main/java/com/eduquest/backend/domain/identity/event/RotateTokenEvent.java
@@ -1,4 +1,4 @@
-package com.eduquest.backend.domain.member.event;
+package com.eduquest.backend.domain.identity.event;
 
 import jakarta.servlet.http.HttpServletResponse;
