+++ src/main/java/com/eduquest/backend/domain/identity/model/enums/RoleType.java
@@ -1,0 +1,7 @@
+package com.eduquest.backend.domain.identity.model.enums;
+
+public enum RoleType {
+    USER,
+    ADMIN
+
+}
