--- src/main/java/com/eduquest/backend/domain/member/model/enums/RoleType.java
@@ -1,7 +1,0 @@
-package com.eduquest.backend.domain.member.model.enums;
-
-public enum RoleType {
-    USER,
-    ADMIN
-
-}
