+++ src/main/java/com/eduquest/backend/infrastructure/persistence/identity/repository/impl/MemberQRepositoryImpl.java
@@ -1,7 +1,7 @@
 package com.eduquest.backend.infrastructure.persistence.identity.repository.impl;
 
-import com.eduquest.backend.domain.member.dto.MemberQuery;
-import com.eduquest.backend.domain.member.dto.UserDetailsData;
+import com.eduquest.backend.domain.identity.dto.MemberQuery;
+import com.eduquest.backend.domain.identity.dto.UserDetailsData;
 import com.eduquest.backend.infrastructure.persistence.identity.entity.QMemberEntity;
 import com.eduquest.backend.infrastructure.persistence.identity.entity.QRoleEntity;
 import com.eduquest.backend.infrastructure.persistence.identity.entity.QUserRoleEntity;
