+++ src/main/java/com/eduquest/backend/infrastructure/persistence/identity/repository/impl/RoleQRepository.java
@@ -1,6 +1,6 @@
 package com.eduquest.backend.infrastructure.persistence.identity.repository.impl;
 
-import com.eduquest.backend.domain.member.dto.MemberQuery;
+import com.eduquest.backend.domain.identity.dto.MemberQuery;
 
 import java.util.List;
 import java.util.Optional;
