+++ src/main/java/com/eduquest/backend/infrastructure/security/encoder/SecurityPasswordEncoder.java
@@ -1,6 +1,6 @@
 package com.eduquest.backend.infrastructure.security.encoder;
 
-import com.eduquest.backend.domain.member.component.CustomPasswordEncoder;
+import com.eduquest.backend.domain.identity.component.CustomPasswordEncoder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
